refactor(addon-docs): rewrite storiesOf to dynamic CSF
